Codewars - Sudoku Solution Validator <4kyu>
https://www.codewars.com/kata/529bf0e9bdf7657179000008

Write a function validSolution that accepts a 2D array representing a Sudoku board, and returns true if it is a valid solution, or false otherwise. The cells of the sudoku board may also contain 0's, which will represent empty cells. Boards containing one or more zeroes are considered to be invalid solutions.