Codewars - Valid Braces <4kyu>
https://www.codewars.com/kata/5277c8a221e209d3f6000b56

Write a function called validBraces that takes a string of braces, and determines if the order of the braces is valid. A string of braces is considered valid if all braces are matched with the correct brace. 

Examples: 
validBraces( "(){}[]" ) => returns true 
validBraces( "(}" ) => returns false 
validBraces( "[(])" ) => returns false 
validBraces( "([{}])" ) => returns true